tools/ui/create_data_tools/cropper_app/requirements.txt
# Minimal runtime deps for cropper_app.
# Install: `pip install -r cropper_app/requirements.txt`

PySide6
pillow

